% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{check_primary_color_is_ok}
\usage{
check_data(string)
}
\arguments{
\item{string}{: a character vector. The color of the squirrels.}
}
\value{
all_colors_OK: a character vector or error message
}
\description{
check_primary_color_is_ok
}
\examples{
check_data( c("Gray", "Gray", "Black"))
#check_data( c("Gray", "Gray","Brown"))
}
