% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_primary_color_is_ok.R
\name{check_primary_color_is_ok}
\alias{check_primary_color_is_ok}
\title{check_primary_color_is_ok}
\usage{
check_primary_color_is_ok(string)
}
\arguments{
\item{string}{: a character vector. The color of the squirrels.}
}
\value{
all_colors_OK: a character vector or error message
}
\description{
check_primary_color_is_ok
}
\examples{

check_primary_color_is_ok( c("Gray", "Gray", "Black"))
#check_primary_color_is_ok( c("Gray", "Gray","Brown"))
}
