% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_squirrel_data_integrity.R
\name{check_squirrel_data_integrity}
\alias{check_squirrel_data_integrity}
\title{check_squirrel_data_integrity}
\usage{
check_squirrel_data_integrity(dataset)
}
\arguments{
\item{dataset}{A dataset.}
}
\value{
a message if everything is OK
}
\description{
Description The function checks that the data contains a primary_fur_color column, and that this column contains only allowed colors
}
\examples{
check_squirrel_data_integrity(dataset = data_act_squirrels)

# create dummy data 
#the_data_bad <- data_act_squirrels 
#the_data_bad<- dplyr::rename(the_data_bad, secondary_color = primary_fur_color )
#check_squirrel_data_integrity(dataset = the_data_bad)


}
