% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_act_squirrels.R
\docType{data}
\name{data_act_squirrels}
\alias{data_act_squirrels}
\title{data_act_squirrels}
\format{
A data frame with 15 rows and 4 variables:
\describe{
\item{ age }{  character }
\item{ primary_fur_color }{  character }
\item{ activity }{  character }
\item{ counts }{  numeric }
}
}
\source{
Generated with check helper because sinew is not the option
}
\usage{
data_act_squirrels
}
\description{
A small act squirrels random dataset.
}
\keyword{datasets}
