% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message_fur_color.R
\name{get_message_fur_color}
\alias{get_message_fur_color}
\title{G a message with the fur color of interest.}
\usage{
get_message_fur_color(primary_fur_color)
}
\arguments{
\item{primary_fur_color}{a character string.}
}
\value{
Nothing. Used for side effect. Outputs a message in the console.
}
\description{
G a message with the fur color of interest.
}
\examples{
get_message_fur_color(primary_fur_color = "Black")
get_message_fur_color(primary_fur_color = "White")
get_message_fur_color(primary_fur_color = "Green")
}
